Fetch most recent commit from GitHub API
Uses the most recent commit of the branch fetched explicitly from the
GitHub API, rather than the branch's tarball URL. This avoids using a
stale cached tarball when commits update the branch in close succession
while still caching tarballs when appropriate (i.e. when the commit is
actually the same).